requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
pypdf
chromadb
langchain
tiktoken
python-multipart
pytest
httpx
